created a reusable  dropdown component
