--- src/components/Testdropdown.css
@@ -1,41 +1,0 @@
-.dd-wrapper {
-  display: flex;
-  min-height: 38px;
-  flex-wrap: wrap;
-}
-.dd-header {
-  display: flex;
-  justify-content: space-between;
-  cursor: pointer;
-  width: 100%;
-  padding: 0 20px;
-}
-
-.dd-list {
-  box-shadow: 0 0.125rem 0.25rem rgba(0, 0, 0, 0.075) !important;
-  padding: 0;
-  margin: 0;
-  width: 100%;
-  margin-top: 20px;
-  list-style: none;
-}
-.dd-list-item {
-  list-style: none;
-}
-
-dd-btn {
-  border-top: 1px solid #ccc;
-  border-top-left-radius: 4px;
-  border-top-right-radius: 4px;
-  display: flex;
-  justify-content: space-between;
-  background-color: white;
-  font-size: 16px;
-  padding: 15px 20px 15px 20px;
-  border: 0;
-  border-bottom: 1px solid #ccc;
-  width: 100%;
-  text-align: left;
-  border-left: 1px solid #ccc;
-  border-right: 1px solid #ccc;
-}
